x-growth-agent/requirements.txt
anthropic>=0.40.0
tweepy>=4.14.0
apscheduler>=3.10.4
numpy>=1.26.0
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.7
